Created separate task executer class(es) #29
